% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aj_estimate_helpers.R
\name{extract_crude_estimate}
\alias{extract_crude_estimate}
\title{Extract crude estimate}
\usage{
extract_crude_estimate(data_to_adjust)
}
\arguments{
\item{data_to_adjust}{A data frame contains original estimated risks,
estimated risks stratas, the time to events, and the stratification technique}
}
\value{
A data frame contains the crude estimates, i.e the counts for each state.
}
\description{
This function extract the crude estimates of the
state occupancy in the end of the followup
}
