% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aj_estimate_helpers.R
\name{update_administrative_censoring}
\alias{update_administrative_censoring}
\title{Update administrative censoring}
\usage{
update_administrative_censoring(data_to_adjust)
}
\arguments{
\item{data_to_adjust}{A data frame contains original estimated risks,
estimated risks stratas, the time to events, and the stratification technique}
}
\value{
A modified version of `data_to_adjust`.
}
\description{
This function update outcomes according to fixed time horizons:
If no event happened until fixed time horizon the outcome will be updated
to `real_negatives`. If the original outcome is already defined as `real_negatives`
but the tiemes is updated prior to the fixed time horizon, the outcome will be
updated as `real_censored`.
}
